requirements.txt
aiogram<3.0
environs~=8.0.0
Pillow==9.1.0